Add Guild membership feature to User entity
The commit introduces a method called `add_to_guild` in the User entity within src/database/entities/user.rs. This function enables to add a user to a specific guild by guild_id. A check will be performed to verify if the specified guild_id does exist, raising an InvalidGuild error if it does not. It also accounts for a scenario where a user could already be in the guild, returning an error in such cases. Future implementations would include checks for limiting maximum guild count and banning certain users.